Move CouchbaseLiteFeedStore to its own file in production
